@@ -15,7 +15,7 @@
 		"mock-api:watch": "tsx watch ./src/mocks/http.ts",
 		"type-check": "tsc --project tsconfig.validation.json",
 		"type-check:cypress": "tsc --project cypress/tsconfig.json",
-		"lint": "eslint . --ext .js,.jsx,.ts,.tsx --max-warnings 0 --report-unused-disable-directives",
+		"lint": "eslint . --max-warnings 0 --report-unused-disable-directives",
 		"lint:fix": "pnpm lint --fix",
 		"prettier-base": "prettier . --ignore-unknown --cache --log-level warn",
 		"format": "pnpm prettier-base --write",
@@ -68,6 +68,8 @@
 		"@chakra-ui/styled-system": "2.9.2",
 		"@chakra-ui/theme-tools": "2.1.2",
 		"@chakra-ui/utils": "2.0.14",
+		"@eslint/eslintrc": "3.2.0",
+		"@eslint/js": "9.16.0",
 		"@mswjs/http-middleware": "0.10.2",
 		"@octokit/openapi-types": "22.2.0",
 		"@tanstack/eslint-plugin-query": "5.62.1",
@@ -91,7 +93,7 @@
 		"@vitest/eslint-plugin": "1.1.14",
 		"cors": "2.8.5",
 		"cypress": "13.16.0",
-		"eslint": "8.57.1",
+		"eslint": "9.16.0",
 		"eslint-config-next": "15.0.3",
 		"eslint-config-prettier": "9.1.0",
 		"eslint-plugin-cypress": "4.1.0",
